langchain/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
python-jose==3.3.0         # For JWT verification
httpx==0.27.0               # Optional: to call external APIs like Gemini/OpenAI
python-dotenv==1.0.1        # For loading .env secrets

# LangChain dependencies
langchain==0.1.11
langchain-google-genai==0.0.9  # Google Generative AI integration
pydantic==2.7.0
